Add tracking camera

+++ Source/Settings.cpp
@@ -3,7 +3,7 @@
 namespace SpaceSim
 {
     double PixelPerM     = 1.36571e-09;
-    double ZoomScale     = .5f;
+    float  ZoomScale     = .5f;
     double TimeScale     = 60.0 * 60.0 * 24.0 * 20.0;
     bool   ToScaleBodies = false;
 } // namespace SpaceSim
